+++ extensions/catalog/federated-catalog-spi/src/main/java/org/eclipse/dataspaceconnector/catalog/spi/Loader.java
@@ -22,7 +22,14 @@ import java.util.Collection;
  * Puts the result of a catalog request (i.e. {@link org.eclipse.dataspaceconnector.catalog.spi.model.UpdateResponse} into
  * whatever storage backend or database is used.
  */
-@FunctionalInterface
 public interface Loader {
+    /**
+     * Stores a list of {@link UpdateResponse}s in its internal database.
+     */
     void load(Collection<UpdateResponse> batch);
+
+    /**
+     * Deletes all stored entries. Typically called before loading a new batch.
+     */
+    void clear();
 }